% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_03_conservation.R
\name{conservation_ui}
\alias{conservation_ui}
\alias{conservation_server}
\title{Conservation module}
\usage{
conservation_ui(id)

conservation_server(id)
}
\arguments{
\item{id}{Module ID, should match across each UI and Server pairing.}
}
\value{
Either the HTML UI components, or the corresponding server code for the module.
}
\description{
Create the UI and Server components for the conservation page.
}
