% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_01_diversity.R
\name{diversity_ui}
\alias{diversity_ui}
\alias{diversity_server}
\title{Diversity module}
\usage{
diversity_ui(id)

diversity_server(id)
}
\arguments{
\item{id}{Module ID, should match across each UI and Server pairing.}
}
\value{
Either the HTML UI components, or the corresponding server code for the module.
}
\description{
Create the UI and Server components for the diversity page.
}
