% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_if_truthy.R
\name{filter_if_truthy}
\alias{filter_if_truthy}
\title{Filter data if a condition \code{isTruthy}}
\usage{
filter_if_truthy(.data, col, ifTruthy)
}
\arguments{
\item{.data}{A dataset to be filtered.}

\item{col}{<\code{\link[rlang:args_data_masking]{data-masking}}> A column for that should be filtered by
for matches against the \code{isTruthy} variable.}

\item{ifTruthy}{A value that is check for Truthyness using \code{\link[shiny:isTruthy]{shiny::isTruthy()}}}
}
\value{
The filtered dataset if the \code{ifTruthy} variable if Truthy, or the original dataset if
not.
}
\description{
Checks if a condition meets the \code{\link[shiny:isTruthy]{shiny::isTruthy()}} condition, and if it is then filter a column
of the data on matches to the \code{ifTruthy} value.
}
\details{
When the condition of \code{ifTruthy} is found to be \code{TRUE} according to \code{\link[shiny:isTruthy]{shiny::isTruthy()}}, then the
data is filtered where \code{col \%in\% ifTruthy}.
}
