% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_04_gbf_indicators.R
\name{gbf_indicators_ui}
\alias{gbf_indicators_ui}
\alias{gbf_indicators_server}
\title{GBF Indicators module}
\usage{
gbf_indicators_ui(id)

gbf_indicators_server(id)
}
\arguments{
\item{id}{Module ID, should match across each UI and Server pairing.}
}
\value{
Either the HTML UI components, or the corresponding server code for the module.
}
\description{
Create the UI and Server components for the risk page.
}
