% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_layer_select.R
\name{layer_select_ui}
\alias{layer_select_ui}
\alias{layer_select_server}
\title{Create layer selection accordion}
\usage{
layer_select_ui(id, spec_file)

layer_select_server(id, spec_file)
}
\arguments{
\item{spec_file}{File path to a YAML file that specifies what options should be available in the
accordion.}
}
\value{
A \code{\link[bslib:accordion]{bslib::accordion()}} with a panel for each top-level entry in the YAML file.

The server part of the module returns a list denoting which Accordion panel is active,
and which layer is selected inside of that panel.
}
\description{
Based on the spec in a yaml file, create an accordion for selecting different layers.
}
