% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_02_risk_ecosystems.R
\name{risk_ecosystem_ui}
\alias{risk_ecosystem_ui}
\alias{risk_ecosystem_server}
\title{Risk ecosystems module}
\usage{
risk_ecosystem_ui(id)

risk_ecosystem_server(id)
}
\arguments{
\item{id}{Module ID, should match across each UI and Server pairing.}
}
\value{
Either the HTML UI components, or the corresponding server code for the module.
}
\description{
Create the UI and Server components for a risk ecosystems page.
}
