% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_02_risk_edge_countries.R,
%   R/mod_02_risk_edge_species.R
\name{risk_edge_countries_ui}
\alias{risk_edge_countries_ui}
\alias{risk_edge_countries_server}
\alias{risk_edge_species_ui}
\alias{risk_edge_species_server}
\title{Risk EDGE module}
\usage{
risk_edge_countries_ui(id)

risk_edge_countries_server(id, edge_countries)

risk_edge_species_ui(id)

risk_edge_species_server(id, edge_countries)
}
\arguments{
\item{id}{Module ID, should match across each UI and Server pairing.}

\item{edge_countries}{Dataset for EDGE countries}
}
\value{
Either the HTML UI components, or the corresponding server code for the module.
}
\description{
Create the UI and Server components for the risk page.
}
