% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_02_risk_redlist.R
\name{risk_redlist_ui}
\alias{risk_redlist_ui}
\alias{risk_redlist_server}
\title{Risk redlist module}
\usage{
risk_redlist_ui(id)

risk_redlist_server(id)
}
\arguments{
\item{id}{Module ID, should match across each UI and Server pairing.}

\item{edge_countries}{Dataset for EDGE countries}
}
\value{
Either the HTML UI components, or the corresponding server code for the module.
}
\description{
Create the UI and Server components for a risk redlist page.
}
