% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_02_risk.R
\name{risk_ui}
\alias{risk_ui}
\alias{risk_server}
\title{Risk module}
\usage{
risk_ui(id)

risk_server(id)
}
\arguments{
\item{id}{Module ID, should match across each UI and Server pairing.}
}
\value{
Either the HTML UI components, or the corresponding server code for the module.
}
\description{
Create the UI and Server components for the risk page.
}
