% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Main application server function}
\usage{
server(input, output, session)
}
\arguments{
\item{input, output}{Shiny reactive values for the app's input and output objects.}

\item{session}{A shiny session object.}
}
\value{
A Shiny server.
}
\description{
Main application server function
}
