% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_datatable.R
\name{tab_datatable_ui}
\alias{tab_datatable_ui}
\alias{tab_datatable_server}
\title{Display table of data as a Nav Panel.}
\usage{
tab_datatable_ui(id, title = "Table", ..., value = title, icon = NULL)

tab_datatable_server(
  id,
  .data,
  options = list(searching = FALSE, pageLength = 5, scrollX = TRUE, scrollY = FALSE, dom
    = "Bftip", lengthChange = FALSE, buttons = list(list(extend = "csv", text =
    "Download CSV", exportOptions = list(modifier = list(modifier = list(page =
    "all"))))))
)
}
\arguments{
\item{id}{Module ID, should match across each UI and Server pairing.}

\item{title}{A title to display. Can be a character string or UI elements
(i.e., \link[htmltools:builder]{htmltools::tags}).}

\item{...}{Depends on the function:
\itemize{
\item For \code{nav_panel()} and \code{nav_panel_hidden()}: UI elements (i.e.,
\link[htmltools:builder]{htmltools::tags}) to display when the item is active.
\item For \code{nav_menu()}: a collection of nav items (e.g., \code{nav_panel()},
\code{nav_item()}).
\item For \code{nav_item()}: UI elements (i.e., \link[htmltools:builder]{htmltools::tags}) to place directly in
the navigation panel (e.g., search forms, links to external content, etc).
}}

\item{value}{A character string to assign to the nav item. This value may be
supplied to the relevant container's \code{selected} argument in order to show
particular nav item's content immediately on page load. This value is also
useful for programmatically updating the selected content via
\code{\link[bslib:nav_select]{nav_select()}}, \code{\link[bslib:nav_hide]{nav_hide()}}, etc (updating selected tabs this way is often
useful for showing/hiding panels of content via other UI controls like
\code{\link[shiny:radioButtons]{shiny::radioButtons()}} -- in this scenario, consider using
\code{\link[bslib:nav_panel_hidden]{nav_panel_hidden()}} with \code{\link[bslib:navset_hidden]{navset_hidden()}}).}

\item{icon}{Optional icon to appear next to the nav item's \code{title}.}

\item{.data}{Data to present in the table, as a reactive expression.}

\item{options}{Additional Datatable options passed to \code{\link[DT:datatable]{DT::datatable()}}.}
}
\value{
A \code{\link[bslib:nav-items]{bslib::nav_panel()}} containing a \code{\link[DT:datatable]{DT::datatable()}} representation of some input
data.

From \code{tab_datatable_server()}, the \code{\link[DT:datatable]{DT::datatable()}} object as a reactive expression.
}
\description{
Creates a \code{\link[bslib:nav-items]{bslib::nav_panel()}} page containing a \code{\link[DT:datatable]{DT::datatable()}} representation of some input
data.
}
