% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui}
\alias{ui}
\title{Main application user interface}
\usage{
ui()
}
\value{
A \code{\link[bslib:page_navbar]{bslib::page_navbar()}} object of the application user interface.
}
\description{
Main application user interface
}
