% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_datatable.R
\name{tab_datatable_ui}
\alias{tab_datatable_ui}
\alias{tab_datatable_server}
\title{Display table of data as a Nav Panel.}
\usage{
tab_datatable_ui(
  id,
  title = "Table",
  ...,
  value = title,
  icon = NULL,
  as_nav_panel = TRUE
)

tab_datatable_server(
  id,
  .data,
  options = list(searching = FALSE, pageLength = 5, scrollX = TRUE, scrollY = FALSE, dom
    = "Bftip", lengthChange = FALSE, buttons = list(list(extend = "csv", text =
    "Download CSV", exportOptions = list(modifier = list(modifier = list(page =
    "all")))))),
  extensions = "Buttons",
  class = "compact stripe hover nowrap",
  selection = c("multiple", "single", "none")
)
}
\arguments{
\item{id}{Module ID, should match across each UI and Server pairing.}

\item{title}{A title to display. Can be a character string or UI elements
(i.e., \link[htmltools:builder]{htmltools::tags}).}

\item{...}{Additional arguments passed to \code{\link[DT:datatable]{DT::datatable()}}.}

\item{value}{A character string to assign to the nav item. This value may be
supplied to the relevant container's \code{selected} argument in order to show
particular nav item's content immediately on page load. This value is also
useful for programmatically updating the selected content via
\code{\link[bslib:nav_select]{nav_select()}}, \code{\link[bslib:nav_hide]{nav_hide()}}, etc (updating selected tabs this way is often
useful for showing/hiding panels of content via other UI controls like
\code{\link[shiny:radioButtons]{shiny::radioButtons()}} -- in this scenario, consider using
\code{\link[bslib:nav_panel_hidden]{nav_panel_hidden()}} with \code{\link[bslib:navset_hidden]{navset_hidden()}}).}

\item{icon}{Optional icon to appear next to the nav item's \code{title}.}

\item{as_nav_panel}{Whether to return the table as a standalone \code{\link[DT:datatable]{DT::datatable()}} (when \code{FALSE}),
or to wrap the table in a \code{\link[bslib:nav-items]{bslib::nav_panel()}} (when \code{TRUE}, by default).}

\item{.data}{Data to present in the table, as a reactive expression.}

\item{options}{a list of initialization options (see
\url{https://datatables.net/reference/option/}); the character options
wrapped in \code{\link[htmlwidgets]{JS}()} will be treated as literal
JavaScript code instead of normal character strings; you can also set
options globally via \code{\link{options}(DT.options = list(...))}, and
global options will be merged into this \code{options} argument if set}

\item{extensions}{a character vector of the names of the DataTables
extensions (\url{https://datatables.net/extensions/index})}

\item{class}{the CSS class(es) of the table; see
\url{https://datatables.net/manual/styling/classes}}

\item{selection}{the row/column selection mode (single or multiple selection
or disable selection) when a table widget is rendered in a Shiny app;
alternatively, you can use a list of the form \code{list(mode = 'multiple',
selected = c(1, 3, 8), target = 'row', selectable = c(-2, -3))} to
pre-select rows and control the selectable range; the element
\code{target} in the list can be \code{'column'} to enable column
selection, or \code{'row+column'} to make it possible to select both rows
and columns (click on the footer to select columns), or \code{'cell'} to
select cells. See details section for more info.}
}
\value{
A \code{\link[bslib:nav-items]{bslib::nav_panel()}} containing a \code{\link[DT:datatable]{DT::datatable()}} representation of some input
data.

From \code{tab_datatable_server()}, the \code{\link[DT:datatable]{DT::datatable()}} object as a reactive expression.
}
\description{
Creates a \code{\link[bslib:nav-items]{bslib::nav_panel()}} page containing a \code{\link[DT:datatable]{DT::datatable()}} representation of some input
data.
}
